+++ .bumpversion.cfg
@@ -1,0 +1,13 @@
+[bumpversion]
+current_version = 0.0.0
+commit = True
+tag = True
+tag_name = {new_version}
+
+[bumpversion:file:pyproject.toml]
+search = version = "{current_version}"
+replace = version = "{new_version}"
+
+[bumpversion:file:src/aiosignalrcore/__init__.py]
+search = __version__ = '{current_version}'
+replace = __version__ = '{new_version}'
