+++ docs/README.md
@@ -191,7 +191,7 @@ hub_connection = HubConnectionBuilder()\
 ## Configure messagepack
 
 ```python
-from signalrcore.protocol.messagepack_protocol import MessagePackHubProtocol
+from aiosignalrcore.protocol.messagepack_protocol import MessagePackHubProtocol
 
 HubConnectionBuilder()\
             .with_url(self.server_url, options={"verify_ssl":False})\
@@ -251,7 +251,7 @@ hub_connection.stream(
 ```
 ## Client side Streaming
 ```python
-from signalrcore.subject import  Subject
+from aiosignalrcore.subject import  Subject
 
 subject = Subject()
 
@@ -280,7 +280,7 @@ A mini example could be something like this:
 ```python
 import logging
 import sys
-from signalrcore.hub_connection_builder import HubConnectionBuilder
+from aiosignalrcore.hub_connection_builder import HubConnectionBuilder
 
 
 def input_with_default(input_text, default_value):
