+++ pyproject.toml
@@ -1,16 +1,31 @@
 [tool]
 [tool.poetry]
-name = "signalrcore"
-version = "0.9.2"
-description = "A Python SignalR Core client(json and messagepack), with invocation auth and two way streaming. Compatible with azure / serverless functions. Also with automatic reconnect and manually reconnect."
+name = "aiosignalrcore"
+version = "0.0.0"
+description = "Async fork of Python SignalR Core client(json and messagepack), with invocation auth and two way streaming. Compatible with azure / serverless functions. Also with automatic reconnect and manually reconnect."
 keywords = ["signalr", "core", "client", "3.1"]
 classifiers = ["Programming Language :: Python :: 3.6"]
-homepage = "https://github.com/mandrewcito/signalrcore"
+homepage = "https://github.com/dipdup-net/aiosignalrcore"
 authors = ["mandrewcito <anbaalo@gmail.com>"]
 readme = "README.md"
+
+packages = [
+    { include = "aiosignalrcore", from = "src" },
+]
 
 [tool.poetry.dependencies]
 python = "^3.7"
 msgpack = "==1.0.2"
 requests = ">=2.22.0"
-websocket-client = "==0.54.0"
+websockets = "^8.1"
+
+[tool.poetry.dev-dependencies]
+pytest = "^3.0"
+pytest-cov = "^2.4"
+bump2version = "^1.0.1"
+pylint = "^2.7.2"
+pylint-exit = "^1.2.0"
+isort = "^5.7.0"
+mypy = "^0.812"
+diff-cover = "^4.2.3"
+black = "^20.8b1"
