--- requirements.txt
@@ -1,3 +1,0 @@
-msgpack==1.0.2
-requests>=2.22.0
-websocket-client==0.54.0
