--- signalrcore.code-workspace
@@ -1,9 +1,0 @@
-{
-	"folders": [
-		{
-			"path": "."
-		}
-	],
-	"settings": {
-	}
-}
