+++ src/aiosignalrcore/hub/auth_hub_connection.py
