+++ src/aiosignalrcore/hub/errors.py
