+++ src/aiosignalrcore/messages/base_message.py
