+++ src/aiosignalrcore/messages/close_message.py
