+++ src/aiosignalrcore/messages/handshake/response.py
