+++ src/aiosignalrcore/messages/message_type.py
