+++ src/aiosignalrcore/messages/stream_invocation_message.py
