+++ src/aiosignalrcore/messages/stream_item_message.py
