+++ src/aiosignalrcore/protocol/json_hub_protocol.py
@@ -5,7 +5,7 @@ from .base_hub_protocol import BaseHubProtocol
 from ..messages.message_type import MessageType
 from json import JSONEncoder
 
-from signalrcore.helpers import Helpers
+from aiosignalrcore.helpers import Helpers
 
 
 class MyEncoder(JSONEncoder):
