+++ src/aiosignalrcore/transport/base_transport.py
