+++ test/configuration_test.py
@@ -1,6 +1,6 @@
 import websocket
 import logging
-from signalrcore.hub_connection_builder import HubConnectionBuilder
+from aiosignalrcore.hub_connection_builder import HubConnectionBuilder
 
 from test.base_test_case import BaseTestCase, Urls
