+++ test/examples/chat_auth_py2.py
@@ -1,7 +1,7 @@
 import requests
 import sys
 sys.path.append("./")
-from signalrcore.hub_connection_builder import HubConnectionBuilder
+from aiosignalrcore.hub_connection_builder import HubConnectionBuilder
 
 
 def print_message(x):
