+++ test/examples/chat_manual_reconnect.py
@@ -1,7 +1,7 @@
 import logging
 import sys
 import time
-from signalrcore.hub_connection_builder import HubConnectionBuilder
+from aiosignalrcore.hub_connection_builder import HubConnectionBuilder
 
 
 def input_with_default(input_text, default_value):
