+++ test/examples/disconnect.py
@@ -5,7 +5,7 @@ import time
 
 sys.path.append("./")
 
-from signalrcore.hub_connection_builder import HubConnectionBuilder
+from aiosignalrcore.hub_connection_builder import HubConnectionBuilder
 
 connection = HubConnectionBuilder()\
     .with_url("wss://localhost:5001/chathub", options={"verify_ssl": False})\
