+++ test/examples/stream.py
@@ -2,7 +2,7 @@
 import time
 import sys
 sys.path.append("./")
-from signalrcore.hub_connection_builder import HubConnectionBuilder
+from aiosignalrcore.hub_connection_builder import HubConnectionBuilder
 import logging
